Create LRUCacheTest.java
LRUCacheTest Class

This Java class is a test implementation for a LevelDB database with a small LRU (Least Recently Used) cache. It demonstrates how to use LevelDB to store and retrieve data, and how to test the performance of the cache.

Here's what the class does:

**Creates a LevelDB database**:  It sets up a new LevelDB instance with a small LRU cache of 1MB and a maximum of 256 open files.

**Inserts data into the database**: It inserts 1000 key-value pairs into the database, with keys and values in the format "keyX" and "valueX", respectively.

**Retrieves data from the database**: It retrieves 100 values from the database, measuring the time taken to do so.

**Accesses the first value in the cache**: It seeks to the first entry in the database and retrieves the key and value, demonstrating how to access data in the cache.


This class is useful for testing the performance and functionality of LevelDB with an LRU cache and can serve as a starting point for building more complex database applications.